+++ src/components/atoms/Checkbox/index.tsx
@@ -8,12 +8,22 @@ type CheckboxProps = {
 export function Checkbox(props: CheckboxProps) {
   const { checked, customStyle } = props;
 
-  return (
-    <Container title="checkbox" $customStyle={customStyle}>
-      <label className="control control--checkbox">
-        <input type="checkbox" className="checked" checked={checked} readOnly />
-        <div className="control__indicator" />
-      </label>
-    </Container>
-  );
+ const handleCheckboxClick = (event: React.MouseEvent<HTMLInputElement>) => {
+   event.stopPropagation();
+ };
+
+ return (
+   <Container title="checkbox" $customStyle={customStyle}>
+     <label className="control control--checkbox">
+       <input
+         type="checkbox"
+         className="checked"
+         checked={checked}
+         readOnly
+         onClick={handleCheckboxClick}
+       />
+       <div className="control__indicator" />
+     </label>
+   </Container>
+ );
 }