+++ src/styles/global.styles.ts
@@ -197,10 +197,22 @@ export default createGlobalStyle<GlobalStyleProps>`
     border: 1px solid gray;
     border-radius: 20px;
   }
-
   
+  .read_only {
+    opacity: .5;
+  }
+
   .editing {
-    outline: 1px solid var(--bright) ;
+    animation: glowing 0.56s infinite alternate ease-in-out;
+  }
+
+  @keyframes glowing {
+    0% {
+      outline: 1px solid var(--ground);
+    }
+    100% {
+      outline: 1px solid #454545;
+    }
   }
 
   button, a, input {